organizing utility functions
