+++ src/components/Board.jsx
@@ -1,6 +1,8 @@
 import { useState, useRef, useEffect } from "react";
 import "./Board.css";
 
+import { isRowFilled, handleKeyPress } from "./BoardUtils";
+
 const Board = () => {
   const numRows = 6;
   const numCols = 5;
@@ -9,6 +11,9 @@ const Board = () => {
 
   const initialBoxLetters = Array(numRows).fill(Array(numCols).fill(""));
   const [boxContents, setBoxContents] = useState(initialBoxLetters);
+  const [word, setWord] = useState("");
+  //if enter is pressed then the word will be set as the current row's word
+  // then it would be checked to see what row it fits into
 
   useEffect(() => {
     // Focus on the first input box when the component mounts
@@ -17,37 +22,30 @@ const Board = () => {
     }
   }, []);
   
+  // Perform actions when the row is filled and Enter is pressed
+  const handleEnter = () => {
+    if (isRowFilled(boxContents, numRows)) {
+      console.log("COCONUTS")
+      console.log("Row is filled");
+    } else {
+      console.log("Row is not filled");
+    }
+  }
+
+  // useEffect(() => {
+  //   console.log(boxContents);
+  // }, [boxContents]);
+  console.log(boxContents);
+  const handleKeyDown = (e, rowIndex, colIndex) => {
+    handleKeyPress(boxContents, setBoxContents, rowIndex, colIndex, inputRef, numCols, e);
+    if (e.key === "Enter") {
+      handleEnter();
+    }
+  };
+
   //create a function that would check if the row is filled
   // if the row is filled then click enter is clickable
   // if enter is clicked then 
-
-
-  const handleKeyPress = (e, rowIndex, colIndex) => {
-    const newBoxLetters = JSON.parse(JSON.stringify(boxContents));
-    const inputChar = e.key.toUpperCase();
-
-
-    // Handle Backspace
-    if (inputChar === "BACKSPACE") {
-      newBoxLetters[rowIndex][colIndex] = ""; // Clear the letter inside the box
-      setBoxContents(newBoxLetters);
-      const prevInput = inputRef.current[rowIndex][colIndex - 1];
-      if (prevInput) prevInput.focus();
-    }
-
-    // Check if input is a letter
-    if (/^[a-zA-Z]$/.test(inputChar)) {
-      newBoxLetters[rowIndex][colIndex] = inputChar;
-      setBoxContents(newBoxLetters);
-
-      // Move focus of the box forward
-      if (colIndex < numCols - 1) {
-        const nextInput = inputRef.current[rowIndex][colIndex + 1];
-        if (nextInput) nextInput.focus();
-      }
-    }
-
-  };
 
   return (
     <div>
@@ -58,7 +56,7 @@ const Board = () => {
             <div
               className="square border-box font-large"
               key={colIndex}
-              onKeyDown={(e) => handleKeyPress(e, rowIndex, colIndex)}
+              onKeyDown={(e) => handleKeyDown(e, rowIndex, colIndex)}
               ref={(el) => {
                 // Create and store a ref for each input box
                 if (!inputRef.current[rowIndex]) {
